[Feature #156111138] Modify all views to use from and persist data to database.
All view methods are modified and updated so that they can store data to database and retrieve for use.
In memory storage data structures are removed from the views.Since the relations used in the database, the views are updated
to incorporate these features where necessary. For instance, retrieval of books involves querying three database relations namely:
Books, BookCategories and Author. The view for retrieving books and other similar views have thus been updated and restructured to
 incoporate data from these relatons.