[Feature #156111138] Modify and restructure testcases.
Testcases are modified to work with new changes to data views and models in the application.
Since the application views are using data from database instead of the initial in-memory data, testcases have been restructured to reflect
these modifications.
Also, documentation strings for the methods in testcases have been edited to fit purpose and be descriptive enough.