Ft return book 156111106 (#33)
* Develop (#31)

* - Increase test coverage by writing by providing more testcases for the endpoints.

* - Add testcases for checking whether book title or author fields are composed of integers.

* Transferring API Description file from Apiary.io

* Alternate title and author to correct book fields.

* [Feature #156111138] Add configurations for postgresql database.

* [Feature #156111138]Re-define/restructure book and user models to persist data in database.

Database models are defined and their relationship established. More fields are added to the models so as to capture necessary data as much as possible.

* [Feature #156111138] Modify all views to use from and persist data to database.

All view methods are modified and updated so that they can store data to database and retrieve for use.
In memory storage data structures are removed from the views.Since the relations used in the database, the views are updated
to incorporate these features where necessary. For instance, retrieval of books involves querying three database relations namely:
Books, BookCategories and Author. The view for retrieving books and other similar views have thus been updated and restructured to
 incoporate data from these relatons.

* [Feature #156111138] Modify and restructure testcases.

Testcases are modified to work with new changes to data views and models in the application.
Since the application views are using data from database instead of the initial in-memory data, testcases have been restructured to reflect
these modifications.
Also, documentation strings for the methods in testcases have been edited to fit purpose and be descriptive enough.

* [Feature #156111138] update project configuration, requirements.txt and other related files.

Configuration file is update to include settings for postgresql database while requirements.txt is update to incorporate
new installed project modules.
Imports are optimised in run.py file while those not used are removed.

[Finishes #156111138]

* Modify .travis.yml to correct database configurations

* Modify .travis.yml to correct database configurations

* Improve overall test coverage for the project.

* Fix failing test on deleting a book.

* Get maximum overall test coverage possible.

* Get maximum overall test coverage possible.

* Correct for book title and author; they are interchanged.

* Create a function to initialize application in __init__.py in application package

* [Feature #156111106] Add test cases for returning a borrowed book.

* [Feature #156111106] Add new file which contains methods for managing book rentals.

* [Feature #156111106] Change database models to refine relationships between related models.
[Finishes #156111106]